introduced IConventionResult
Not terribly convinced yet I prefer this approach, but it's an experiment.
This allows us to remove static factory method call and we can, if needed, treat this object as a context object if needed.

Also changed the Execute method on the convention to be void... it technically makes things simpler, although on the other hand, doesn't in any way indicate you're supposed to call result.Something and that you're supposed to do it only once...

We'll just need to play with it and see where it takes us