renames, renames everywhere
So we're shifting the balance a bit here.
ConventionResult is really becoming a conventionContext which will act as a context object gluing together all various pieces of the pipeline, isolating all the statics etc. Kind of like HttpContext or OperationContext in WCF

we're changing the interface to IConventionResultContext. While it's not the best name it's the best I could come up with, without calling it factory (which it's not, since it doesn't return anything) or calling it provider which is just wrong